@@ -1375,9 +1375,13 @@ func TestResourceDVEnrollment(t *testing.T) {
 				DNSNameSettings: &cps.DNSNameSettings{
 					CloneDNSNames: false,
 				},
-				Geography:     "core",
-				SecureNetwork: "enhanced-tls",
-				SNIOnly:       true,
+				Geography:        "core",
+				MustHaveCiphers:  "ak-akamai-2020q1",
+				OCSPStapling:     "on",
+				PreferredCiphers: "ak-akamai-2020q1",
+				QuicEnabled:      false,
+				SecureNetwork:    "enhanced-tls",
+				SNIOnly:          true,
 			},
 			Org: &cps.Org{
 				AddressLineOne: "150 Broadway",
@@ -1597,9 +1601,13 @@ func TestResourceDVEnrollment(t *testing.T) {
 				DNSNameSettings: &cps.DNSNameSettings{
 					CloneDNSNames: false,
 				},
-				Geography:     "core",
-				SecureNetwork: "enhanced-tls",
-				SNIOnly:       true,
+				Geography:        "core",
+				MustHaveCiphers:  "ak-akamai-2020q1",
+				OCSPStapling:     "on",
+				PreferredCiphers: "ak-akamai-2020q1",
+				QuicEnabled:      false,
+				SecureNetwork:    "enhanced-tls",
+				SNIOnly:          true,
 			},
 			Org: &cps.Org{
 				AddressLineOne: "150 Broadway",
@@ -1975,9 +1983,13 @@ func TestResourceDVEnrollment(t *testing.T) {
 				DNSNameSettings: &cps.DNSNameSettings{
 					CloneDNSNames: false,
 				},
-				Geography:     "core",
-				SecureNetwork: "enhanced-tls",
-				SNIOnly:       true,
+				Geography:        "core",
+				MustHaveCiphers:  "ak-akamai-2020q1",
+				OCSPStapling:     "on",
+				PreferredCiphers: "ak-akamai-2020q1",
+				QuicEnabled:      false,
+				SecureNetwork:    "enhanced-tls",
+				SNIOnly:          true,
 			},
 			Org: &cps.Org{
 				AddressLineOne: "150 Broadway",
@@ -2097,9 +2109,13 @@ func TestResourceDVEnrollment(t *testing.T) {
 				DNSNameSettings: &cps.DNSNameSettings{
 					CloneDNSNames: false,
 				},
-				Geography:     "core",
-				SecureNetwork: "enhanced-tls",
-				SNIOnly:       true,
+				Geography:        "core",
+				MustHaveCiphers:  "ak-akamai-2020q1",
+				OCSPStapling:     "on",
+				PreferredCiphers: "ak-akamai-2020q1",
+				QuicEnabled:      false,
+				SecureNetwork:    "enhanced-tls",
+				SNIOnly:          true,
 			},
 			Org: &cps.Org{
 				AddressLineOne: "150 Broadway",
@@ -2185,9 +2201,13 @@ func TestResourceDVEnrollmentImport(t *testing.T) {
 				DNSNameSettings: &cps.DNSNameSettings{
 					CloneDNSNames: false,
 				},
-				Geography:     "core",
-				SecureNetwork: "enhanced-tls",
-				SNIOnly:       true,
+				Geography:        "core",
+				MustHaveCiphers:  "ak-akamai-2020q1",
+				OCSPStapling:     "on",
+				PreferredCiphers: "ak-akamai-2020q1",
+				QuicEnabled:      false,
+				SecureNetwork:    "enhanced-tls",
+				SNIOnly:          true,
 			},
 			Org: &cps.Org{
 				AddressLineOne: "150 Broadway",
