@@ -44,7 +44,7 @@ func resourceCPSThirdPartyEnrollment() *schema.Resource {
 				Type:        schema.TypeBool,
 				Optional:    true,
 				Default:     false,
-				Description: "Allow to duplicate common name",
+				Description: "Allow to duplicate common name. Default is false",
 			},
 			"sans": {
 				Type:        schema.TypeSet,
@@ -68,7 +68,7 @@ func resourceCPSThirdPartyEnrollment() *schema.Resource {
 				Type:        schema.TypeBool,
 				Optional:    true,
 				Default:     false,
-				Description: "Whether acknowledge warnings before certificate verification",
+				Description: "Whether acknowledge warnings before certificate verification. Default is false",
 			},
 			"auto_approve_warnings": {
 				Type:        schema.TypeSet,
@@ -88,7 +88,7 @@ func resourceCPSThirdPartyEnrollment() *schema.Resource {
 				Type:        schema.TypeString,
 				Default:     "default",
 				Optional:    true,
-				Description: "Certificate trust chain type",
+				Description: "Certificate trust chain type. Default is 'default'",
 			},
 			"csr": {
 				Type:        schema.TypeSet,
