@@ -75,7 +75,7 @@ func TestDataGTMGeoMaps(t *testing.T) {
 							Href: "TestHref2",
 						}},
 					},
-				}, nil, 3)
+				}, nil, testutils.ThreeTimes)
 			},
 			expectedAttributes: map[string]string{
 				"domain":                                 "gtm_terra_testdomain.akadns.net",
@@ -111,7 +111,7 @@ func TestDataGTMGeoMaps(t *testing.T) {
 		"error response from api": {
 			givenTF: "valid.tf",
 			init: func(m *gtm.Mock) {
-				mockListGeoMaps(m, nil, fmt.Errorf("API error"), 1)
+				mockListGeoMaps(m, nil, fmt.Errorf("API error"), testutils.Once)
 			},
 			expectError: regexp.MustCompile("API error"),
 		},
@@ -129,7 +129,7 @@ func TestDataGTMGeoMaps(t *testing.T) {
 						Rel:  "TestRel",
 						Href: "TestHref",
 					}},
-				}}, nil, 3)
+				}}, nil, testutils.ThreeTimes)
 			},
 			expectedAttributes: map[string]string{
 				"domain":          "gtm_terra_testdomain.akadns.net",
