@@ -21,22 +21,22 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("create CIDRMap", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		mockCreateCIDRMap(client, getCIDRMap(), &gtm.CreateCIDRMapResponse{
 			Resource: getCIDRMap(),
 			Status:   getDefaultResponseStatus(),
 		}, nil)
 
-		mockGetCIDRMap(client, getCIDRMap(), nil, 4)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.FourTimes)
 
-		mockGetDatacenter(client, datacenterID5400, &dc, nil, 1)
+		mockGetDatacenter(client, datacenterID5400, &dc, nil, testutils.Once)
 
-		mockGetDomainStatus(client, 2)
+		mockGetDomainStatus(client, testutils.Twice)
 
 		mockUpdateCIDRMap(client, &gtm.UpdateCIDRMapResponse{Status: getDefaultResponseStatus()}, nil)
 
-		mockGetCIDRMap(client, getCIDRMapUpdated(), nil, 3)
+		mockGetCIDRMap(client, getCIDRMapUpdated(), nil, testutils.ThreeTimes)
 
 		mockDeleteCIDRMap(client)
 
@@ -68,18 +68,18 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("update CIDRMap failed", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		mockCreateCIDRMap(client, getCIDRMap(), &gtm.CreateCIDRMapResponse{
 			Resource: getCIDRMap(),
 			Status:   getDefaultResponseStatus(),
 		}, nil)
 
-		mockGetCIDRMap(client, getCIDRMap(), nil, 4)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.FourTimes)
 
-		mockGetDatacenter(client, datacenterID5400, &dc, nil, 1)
+		mockGetDatacenter(client, datacenterID5400, &dc, nil, testutils.Once)
 
-		mockGetDomainStatus(client, 1)
+		mockGetDomainStatus(client, testutils.Once)
 
 		mockUpdateCIDRMap(client, nil, &gtm.Error{
 			Type:       "internal_error",
@@ -88,7 +88,7 @@ func TestResGTMCIDRMap(t *testing.T) {
 			StatusCode: http.StatusInternalServerError,
 		})
 
-		mockGetCIDRMap(client, getCIDRMap(), nil, 1)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.Once)
 
 		mockDeleteCIDRMap(client)
 
@@ -118,22 +118,22 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("create cidrmap, remove outside of terraform, expect non-empty plan", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		mockCreateCIDRMap(client, getCIDRMap(), &gtm.CreateCIDRMapResponse{
 			Resource: getCIDRMap(),
 			Status:   getDefaultResponseStatus(),
 		}, nil)
 
-		mockGetCIDRMap(client, getCIDRMap(), nil, 2)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.Twice)
 
-		mockGetDatacenter(client, datacenterID5400, &dc, nil, 1)
+		mockGetDatacenter(client, datacenterID5400, &dc, nil, testutils.Once)
 
 		// Mock that the CIDRMap was deleted outside terraform
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		// For terraform test framework, we need to mock GetCIDRMap as it would actually exist before deletion
-		mockGetCIDRMap(client, getCIDRMap(), nil, 1)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.Once)
 
 		mockDeleteCIDRMap(client)
 
@@ -164,11 +164,11 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("create cidrmap failed", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		mockCreateCIDRMap(client, getCIDRMap(), nil, &gtm.Error{StatusCode: http.StatusBadRequest})
 
-		mockGetDatacenter(client, datacenterID5400, &dc, nil, 1)
+		mockGetDatacenter(client, datacenterID5400, &dc, nil, testutils.Once)
 
 		useClient(client, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -188,7 +188,7 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("create cidrmap failed - cidrmap already exists", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, getCIDRMap(), nil, 1)
+		mockGetCIDRMap(client, getCIDRMap(), nil, testutils.Once)
 
 		useClient(client, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -208,14 +208,14 @@ func TestResGTMCIDRMap(t *testing.T) {
 	t.Run("create cidrmap denied", func(t *testing.T) {
 		client := &gtm.Mock{}
 
-		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+		mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 		mockCreateCIDRMap(client, getCIDRMap(), &gtm.CreateCIDRMapResponse{
 			Resource: getCIDRMap(),
 			Status:   getDeniedResponseStatus(),
 		}, nil)
 
-		mockGetDatacenter(client, datacenterID5400, &dc, nil, 1)
+		mockGetDatacenter(client, datacenterID5400, &dc, nil, testutils.Once)
 
 		useClient(client, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -323,7 +323,7 @@ func TestResGTMCIDRMapImport(t *testing.T) {
 			mapName:    testCIDRMapName,
 			init: func(m *gtm.Mock) {
 				// Read
-				mockGetCIDRMap(m, getImportedCIDRMap(), nil, 2)
+				mockGetCIDRMap(m, getImportedCIDRMap(), nil, testutils.Twice)
 			},
 			stateCheck: test.NewImportChecker().
 				CheckEqual("domain", "gtm_terra_testdomain.akadns.net").
@@ -350,7 +350,7 @@ func TestResGTMCIDRMapImport(t *testing.T) {
 			mapName:    testCIDRMapName,
 			init: func(m *gtm.Mock) {
 				// Read - error
-				mockGetCIDRMap(m, nil, fmt.Errorf("get failed"), 1)
+				mockGetCIDRMap(m, nil, fmt.Errorf("get failed"), testutils.Once)
 			},
 			expectError: regexp.MustCompile(`get failed`),
 		},
@@ -386,18 +386,18 @@ func TestResGTMCIDRMapImport(t *testing.T) {
 func getCIDRMapMocks() *gtm.Mock {
 	client := &gtm.Mock{}
 
-	mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, 1)
+	mockGetCIDRMap(client, nil, &gtm.Error{StatusCode: http.StatusNotFound}, testutils.Once)
 
 	mockCreateCIDRMap(client, getCIDRMapForOrder(), &gtm.CreateCIDRMapResponse{
 		Resource: getCIDRMapForOrderResp(),
 		Status:   getDefaultResponseStatus(),
 	}, nil)
 
-	mockGetCIDRMap(client, getCIDRMapForOrderResp(), nil, 4)
+	mockGetCIDRMap(client, getCIDRMapForOrderResp(), nil, testutils.FourTimes)
 
-	mockGetDatacenter(client, datacenterID5400, getTestDatacenterResp(), nil, 1)
+	mockGetDatacenter(client, datacenterID5400, getTestDatacenterResp(), nil, testutils.Once)
 
-	mockGetDomainStatus(client, 2)
+	mockGetDomainStatus(client, testutils.Twice)
 
 	mockDeleteCIDRMap(client)
 
