@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/papi"
+	"github.com/akamai/terraform-provider-akamai/v7/pkg/common/ptr"
 	"github.com/akamai/terraform-provider-akamai/v7/pkg/common/testutils"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
@@ -18,7 +19,7 @@ func TestDataProperties(t *testing.T) {
 		client.On("GetProperties",
 			testutils.MockContext,
 			papi.GetPropertiesRequest{GroupID: "grp_test", ContractID: "ctr_test"},
-		).Return(&papi.GetPropertiesResponse{Properties: props}, nil)
+		).Return(&papi.GetPropertiesResponse{Properties: props}, nil).Times(3)
 
 		useClient(client, nil, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -41,7 +42,7 @@ func TestDataProperties(t *testing.T) {
 		client.On("GetProperties",
 			testutils.MockContext,
 			papi.GetPropertiesRequest{GroupID: "grp_test", ContractID: "ctr_test"},
-		).Return(&papi.GetPropertiesResponse{Properties: props}, nil)
+		).Return(&papi.GetPropertiesResponse{Properties: props}, nil).Times(3)
 
 		useClient(client, nil, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -64,7 +65,7 @@ func TestDataProperties(t *testing.T) {
 		client.On("GetProperties",
 			testutils.MockContext,
 			papi.GetPropertiesRequest{GroupID: "grp_test", ContractID: "ctr_test"},
-		).Return(&papi.GetPropertiesResponse{Properties: props}, nil)
+		).Return(&papi.GetPropertiesResponse{Properties: props}, nil).Times(3)
 
 		useClient(client, nil, func() {
 			resource.UnitTest(t, resource.TestCase{
@@ -78,22 +79,76 @@ func TestDataProperties(t *testing.T) {
 
 		client.AssertExpectations(t)
 	})
+
+	t.Run("list properties, in some hostname bucket support enabled", func(t *testing.T) {
+		client := &papi.Mock{}
+		props := papi.PropertiesItems{Items: []*papi.Property{
+			{
+				AccountID:     "act1",
+				AssetID:       "ast1",
+				ContractID:    "ctr_test",
+				GroupID:       "grp_test",
+				LatestVersion: 1,
+				Note:          "note1",
+				PropertyID:    "prp1",
+				PropertyName:  "prpname1",
+				PropertyType:  ptr.To("HOSTNAME_BUCKET"),
+			},
+			{
+				AccountID:     "act2",
+				AssetID:       "ast2",
+				ContractID:    "ctr_test",
+				GroupID:       "grp_test",
+				LatestVersion: 1,
+				Note:          "note2",
+				PropertyID:    "prp2",
+				PropertyName:  "prpname2",
+			},
+			{
+				AccountID:     "act3",
+				AssetID:       "ast3",
+				ContractID:    "ctr_test",
+				GroupID:       "grp_test",
+				LatestVersion: 1,
+				Note:          "note3",
+				PropertyID:    "prp3",
+				PropertyName:  "prpname3",
+				PropertyType:  ptr.To("HOSTNAME_BUCKET"),
+			},
+		}}
+		properties := decodePropertyItems(props.Items)
+
+		client.On("GetProperties",
+			testutils.MockContext,
+			papi.GetPropertiesRequest{GroupID: "grp_test", ContractID: "ctr_test"},
+		).Return(&papi.GetPropertiesResponse{Properties: props}, nil).Times(3)
+
+		useClient(client, nil, func() {
+			resource.UnitTest(t, resource.TestCase{
+				ProtoV6ProviderFactories: testutils.NewProtoV6ProviderFactory(NewSubprovider()),
+				Steps: []resource.TestStep{{
+					Config: testutils.LoadFixtureString(t, "testdata/TestDataProperties/properties.tf"),
+					Check:  buildAggregatedTest(properties, "grp_testctr_test", "grp_test", "ctr_test"),
+				}},
+			})
+		})
+
+		client.AssertExpectations(t)
+	})
 }
 
 func buildPapiProperties() []*papi.Property {
 	properties := make([]*papi.Property, 10)
 	for i := 0; i < 10; i++ {
 		properties[i] = &papi.Property{
-			AccountID:         fmt.Sprintf("act%v", i),
-			AssetID:           fmt.Sprintf("ast%v", i),
-			ContractID:        "ctr_test",
-			GroupID:           "grp_test",
-			LatestVersion:     1,
-			Note:              fmt.Sprintf("note%v", i),
-			ProductionVersion: nil,
-			PropertyID:        fmt.Sprintf("prp%v", i),
-			PropertyName:      fmt.Sprintf("prpname%v", i),
-			StagingVersion:    nil,
+			AccountID:     fmt.Sprintf("act%v", i),
+			AssetID:       fmt.Sprintf("ast%v", i),
+			ContractID:    "ctr_test",
+			GroupID:       "grp_test",
+			LatestVersion: 1,
+			Note:          fmt.Sprintf("note%v", i),
+			PropertyID:    fmt.Sprintf("prp%v", i),
+			PropertyName:  fmt.Sprintf("prpname%v", i),
 		}
 	}
 	return properties
@@ -130,6 +185,11 @@ func decodePropertyItems(items []*papi.Property) []map[string]interface{} {
 			"property_id":        item.PropertyID,
 			"property_name":      item.PropertyName,
 			"staging_version":    decodeVersion(item.StagingVersion),
+		}
+		if item.PropertyType != nil {
+			prop["property_type"] = *item.PropertyType
+		} else {
+			prop["property_type"] = ""
 		}
 		properties = append(properties, prop)
 	}
