@@ -65,9 +65,11 @@ var (
 	// PAPI CP Code errors
 
 	// ErrLookingUpCPCode is returned when fetching CP Code from API client by contractID returned an error or no CP Code was found
-	ErrLookingUpCPCode = errors.New("looking up CP Code by name")
-	// ErrCpCodeNotFound is returned when cp code with provided ID does not exist
-	ErrCpCodeNotFound = errors.New("cp code not found")
+	ErrLookingUpCPCode = errors.New("looking up cp code by name")
+	// ErrCPCodeNotFound is returned when cp code with provided ID does not exist
+	ErrCPCodeNotFound = errors.New("cp code not found")
+	// ErrMoreCPCodesFound is returned when cp code with provided ID does not exist
+	ErrMoreCPCodesFound = errors.New("more cp codes found")
 	// ErrCPCodeUpdateTimeout is returned when waiting for a cp code update results in timeout
 	ErrCPCodeUpdateTimeout = errors.New("cp code update timeout")
 
